refactor: updated territory stories folders and namespaces
