+++ src/Application/Interfaces/Services/TerritoryStories/ITerritoryStoryService.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services.TerritoryStory;
+﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services.TerritoryStories;
 
 using System.Threading;
 using System.Threading.Tasks;
@@ -7,12 +7,13 @@ using IAVH.BioTablero.CM.Application.DTOs.TerritoryStories;
 using IAVH.BioTablero.CM.Application.Interfaces.General;
 using IAVH.BioTablero.CM.Application.Utils;
 using IAVH.BioTablero.CM.Core.Domain.Entities.TerritoryStories;
-using IAVH.BioTablero.CM.Core.Interfaces.ExternalServices;
+
+using Microsoft.AspNetCore.OData.Query;
 
 /// <summary>
-/// Territory Story Image service interface.
+/// Territory Story service interface.
 /// </summary>
-public interface ITerritoryStoryImageService : IRead<TerritoryStoryImage, int>
+public interface ITerritoryStoryService : IRead<TerritoryStory, int>, IAdd<TerritoryStoryDto>
 {
     /// <summary>
     /// Get element.
@@ -24,22 +25,14 @@ public interface ITerritoryStoryImageService : IRead<TerritoryStoryImage, int>
     Task<CustomWebResponse> GetItemAsync(int id, string userName, CancellationToken ct = default);
 
     /// <summary>
-    /// Get elements by territory story.
+    /// Get entities by initiative (OData).
     /// </summary>
-    /// <param name="territoryStoryId">Territory Story identifier.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>Entities by selected territory story.</returns>
-    Task<CustomWebResponse> GetByTerritoryStoryAsync(int territoryStoryId, CancellationToken ct = default);
-
-    /// <summary>
-    /// Add element.
-    /// </summary>
+    /// <param name="initiativeId">Initiative identifier.</param>
     /// <param name="userName">User name.</param>
-    /// <param name="entityData">Entity data.</param>
-    /// <param name="formFile">Image data.</param>
+    /// <param name="queryOptions">OData query options.</param>
     /// <param name="ct">Cancellation token.</param>
     /// <returns>Process result.</returns>
-    Task<CustomWebResponse> AddAsync(string userName, TerritoryStoryImageDto entityData, IInputFile formFile, CancellationToken ct = default);
+    Task<CustomWebResponse> GetByInitiativeAsync(int initiativeId, string userName, ODataQueryOptions<TerritoryStory> queryOptions, CancellationToken ct = default);
 
     /// <summary>
     /// Update element.
@@ -47,10 +40,17 @@ public interface ITerritoryStoryImageService : IRead<TerritoryStoryImage, int>
     /// <param name="id">Element identifier.</param>
     /// <param name="userName">User name.</param>
     /// <param name="entityData">Entity data.</param>
-    /// <param name="formFile">Image data.</param>
     /// <param name="ct">Cancellation token.</param>
     /// <returns>Process result.</returns>
-    Task<CustomWebResponse> UpdateAsync(int id, string userName, TerritoryStoryImageDto entityData, IInputFile formFile, CancellationToken ct = default);
+    Task<CustomWebResponse> UpdateAsync(int id, string userName, TerritoryStoryDto entityData, CancellationToken ct = default);
+
+    /// <summary>
+    /// Like action.
+    /// </summary>
+    /// <param name="entityData">Entity data.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Process result.</returns>
+    Task<CustomWebResponse> LikeActionAsync(TerritoryStoryLikeDto entityData, CancellationToken ct = default);
 
     /// <summary>
     /// Featured content action.
@@ -62,11 +62,20 @@ public interface ITerritoryStoryImageService : IRead<TerritoryStoryImage, int>
     Task<CustomWebResponse> FeaturedContentActionAsync(int id, string userName, CancellationToken ct = default);
 
     /// <summary>
-    /// Delete element.
+    /// Enable element.
     /// </summary>
     /// <param name="id">Element identifier.</param>
     /// <param name="userName">User name.</param>
     /// <param name="ct">Cancellation token.</param>
     /// <returns>Process result.</returns>
-    Task<CustomWebResponse> DeleteAsync(int id, string userName, CancellationToken ct = default);
+    Task<CustomWebResponse> EnableAsync(int id, string userName, CancellationToken ct = default);
+
+    /// <summary>
+    /// Disable element.
+    /// </summary>
+    /// <param name="id">Element identifier.</param>
+    /// <param name="userName">User name.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Process result.</returns>
+    Task<CustomWebResponse> DisableAsync(int id, string userName, CancellationToken ct = default);
 }
