+++ src/Application/Mappings/TerritoryStories/TerritoryStoryImageMappings.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Mappings.TerritoryStory;
+﻿namespace IAVH.BioTablero.CM.Application.Mappings.TerritoryStories;
 
 using System;
 
@@ -7,12 +7,12 @@ using IAVH.BioTablero.CM.Application.Interfaces.General;
 using IAVH.BioTablero.CM.Core.Domain.Entities.TerritoryStories;
 
 /// <summary>
-/// Territory Story Video mappings.
+/// Territory Story Image mappings.
 /// </summary>
-public class TerritoryStoryVideoMappings : IMapper<TerritoryStoryVideo, TerritoryStoryVideoDto>
+public class TerritoryStoryImageMappings : IMapper<TerritoryStoryImage, TerritoryStoryImageDto>
 {
     /// <inheritdoc/>
-    public TerritoryStoryVideoDto Map(TerritoryStoryVideo entity)
+    public TerritoryStoryImageDto Map(TerritoryStoryImage entity)
     {
         ArgumentNullException.ThrowIfNull(entity);
 
@@ -20,12 +20,14 @@ public class TerritoryStoryVideoMappings : IMapper<TerritoryStoryVideo, Territor
         {
             Id = entity.Id,
             TerritoryStoryId = entity.TerritoryStoryId,
-            FileUrl = entity.FileUrl.ToString(),
+            FileUrl = entity.FileUrl,
+            Description = entity.Description,
+            FeaturedContent = entity.FeaturedContent,
         };
     }
 
     /// <inheritdoc/>
-    public TerritoryStoryVideo Map(TerritoryStoryVideoDto dto)
+    public TerritoryStoryImage Map(TerritoryStoryImageDto dto)
     {
         ArgumentNullException.ThrowIfNull(dto);
 
@@ -33,7 +35,9 @@ public class TerritoryStoryVideoMappings : IMapper<TerritoryStoryVideo, Territor
         {
             Id = dto.Id ?? 0,
             TerritoryStoryId = dto.TerritoryStoryId ?? 0,
-            FileUrl = new Uri(dto.FileUrl),
+            FileUrl = dto.FileUrl,
+            Description = dto.Description,
+            FeaturedContent = dto.FeaturedContent,
         };
     }
 }
